requirements.txt
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.2
docker-pycreds==0.4.0
entrypoints==0.4
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fschat==0.2.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gradio==3.23.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
lit==16.0.2
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.8.10
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pathtools==0.1.2
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
rich==13.3.5
semantic-version==2.10.0
sentencepiece==0.1.98
sentry-sdk==1.21.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
svgwrite==1.4.3
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.15.0
wavedrom==2.0.3.post3
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
